+++ src/redturtle/volto/restapi/services/sitemap/get.py
@@ -3,6 +3,13 @@ from plone.restapi.services import Service
 from plone import api
 
 
+class SitemapSettingsGet(Service):
+    def reply(self):
+        return {"depth": api.portal.get_registry_record("plone.sitemap_depth")}
+
+
+# TODO: add get service for sitemap xml endpoint (?)
+# mimic logic from plone.app.layout.sitemap.sitemap or redturtle.volto.browser.sitemap
 class SitemapGet(Service):
     def reply(self):
         return {"depth": api.portal.get_registry_record("plone.sitemap_depth")}